requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.4.0
pydantic==2.11.4
httpx==0.28.1
starlette==0.36.3
typing_extensions==4.13.2
jinja2==3.1.3
sqlalchemy==2.0.36
aiosqlite==0.20.0